readme
Mom’s Kitchen website is a part of e-commerce. Thus if we own a place we need to upload a menu online to attract potential customers. The Mom’s Kitchen website gives the ability to increase sales and expand their business by giving customers the facility to order food online. With an online menu ordering system, customers can place orders online 24 *7. Thus it is a simple, fast, and convenient food ordering system giving an edge over the competition at an affordable price. With this system, we can set up we menu online and the customers can easily place an order with a simple mouse click.
Also with a food menu online we can easily track the orders, maintain the customer's databases and improve the food delivery service. We can receive orders through the website directly view on the internet. The mom’s kitchen website can even customize the online menu and upload images easily. Having re menu on internet, potential customers can easily access it and place an order at their convenience.